Create helloworld2.json
Same as helloworld.json, except that I ticked the "Embed in map" when adding a new PNG tileset in the Tilesets editor (making it so that it doesn't require me to do a Save As for each tileset (only an OK button).  This will hopefully fix the File Not Found error when loading the map on the rc3 WA test server.