Update helloworld2.json
Adding a Custom Properties to the "collides_layer" ("collides", with type "bool", with the checkbox ticked) didn't make the colliding function.  For this version I edited the Tileset by clicking "Edit Tileset" buton in the Tilesets window, and then gave the BLOCK tile a Custom Property of "collides" with type "bool" with the checkbox ticked.